PRO-7870: Update functional tests config files and add new classes
