+++ build.gradle
@@ -101,9 +101,7 @@ sourceSets {
 task functional(type: Test) {
     description = "Runs functional Tests"
     testClassesDirs = sourceSets.testFunctional.output.classesDirs
-    testClassesDirs += sourceSets.testFunctional.output.classesDirs
     classpath = sourceSets.testFunctional.runtimeClasspath
-    classpath += sourceSets.testFunctional.runtimeClasspath
     finalizedBy aggregate
 }
