+++ src/functionalTest/java/uk/gov/hmcts/probate/functional/cases/AccessCasesTests.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.probate.functional.cases;
+
+public class AccessCasesTests {
+}
