+++ src/functionalTest/java/uk/gov/hmcts/probate/functional/model/Role.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.probate.functional;
+package uk.gov.hmcts.probate.functional.model;
 
 import lombok.Builder;
 import lombok.Data;
@@ -6,6 +6,5 @@ import lombok.Data;
 @Data
 @Builder
 public class Role {
-
     private String code;
 }
