+++ src/functionalTest/java/uk/gov/hmcts/probate/functional/submissions/SubmitDraftCaseTests.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.probate.functional.submissions;
+
+public class SubmitDraftCaseTests {
+}
