+++ src/functionalTest/resources/application.properties
@@ -1,16 +1,15 @@
 probate.submit.url=${TEST_URL:http://localhost:8181}
 probate.business.url=${TEST_BUSINESS_URL:http://localhost:8080}
-probate.persistence.url=${TEST_PERSISTENCE_URL:http://localhost:8282}
 
 ccd.data.store.api.url=http://localhost:4452
 service.auth.provider.base.url=${SERVICE_AUTH_PROVIDER:http://localhost:4502}
-user.auth.provider.oauth2.url=http://localhost:4501
+user.auth.provider.oauth2.url=http://localhost:5000
 service.name=PROBATE_BACKEND
-idam.oauth2.client.secret=123456
-idam.oauth2.client.id=probate
+idam.oauth2.client.secret=ccd_gateway_secret
+idam.oauth2.client.id=ccd_gateway
 idam.oauth2.redirect_uri=http://localhost:3451/oauth2redirect
 env=saat
-idam.secret=${IDAM_SECRET:123456}
+idam.secret=${IDAM_SECRET:ccd_gateway_secret}
 s2s.auth.totp.secret=${S2S_AUTH_TOTP_SECRET:AAAAAAAAAAAAAAAA}
 idam.username=${IDAM_USERNAME:testusername1@test.com}
 idam.userpassword=${IDAM_USERPASSWORD:password}
